% Auto-generated: do not edit by hand
\name{''DragDash}

\alias{''DragDash}

\title{DragDash component}

\description{

}

\usage{
''DragDash(children=NULL, id=NULL, axis=NULL, defaultPosition=NULL,
deltaX=NULL, deltaY=NULL, disabled=NULL, grid=NULL,
handle=NULL, lastX=NULL, lastY=NULL, onStop=NULL,
position=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{axis}{Character. }

\item{defaultPosition}{Named list. }

\item{deltaX}{Numeric. }

\item{deltaY}{Numeric. }

\item{disabled}{Logical. }

\item{grid}{Unnamed list. }

\item{handle}{Character. }

\item{lastX}{Numeric. }

\item{lastY}{Numeric. }

\item{onStop}{}

\item{position}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

